[OR1K] Always address CSRs from FP in stack frames with realignment.
Let's consider the following IR:

  define internal void @foo() alignstack(16) {
    call void asm "", "~{r14}"()
    ret void
  }

Before this change, we emitted prologues with stack realignment
like this:

          .type   foo,@function
  foo:                                    # @foo
          .cfi_startproc
          l.sw    -4(r1), r2
          .cfi_offset 2, -4
          l.addi  r2, r1, 0
          .cfi_def_cfa_register 2
          l.addi  r13, r1, -12
          l.srli  r13, r13, 4
          l.slli  r1, r13, 4
          l.sw    4(r1), r14
          .cfi_offset 14, -8

As you can see, CFI specifies that r14 is saved at r2(CFA)-8.
However, it is actually saved at r1+4. When no realigning is to be
done, r1=r2-12, and r14 is saved at r2-8=(r1+12)-8=r1+4, which is
consistent with CFI. However, if realigning ends up adjusting
the stack, the actual address will be lower than the address in CFI
and attempting to unwind will lead to interesting results.

After this change, the last two instructions generated are:

          l.sw    -8(r2), r14
          .cfi_offset 14, -8

This means that some stack frame objects (CSRs) are addressed one
way (FP-relative and unaligned), and the rest are addressed in
another way (SP-relative and aligned). This is safe, because
the stack is only adjusted down, creating a hole between saved
CSRs and the rest of the frame object, and there are no stack
objects above saved CSRs that could overlap them (RA, FP and BP
are saved above CSRs, but they are also saved before realignment).

The X86 backend does the same:

        .type   foo,@function
foo:                                    # @foo
        .cfi_startproc
        pushq   %rbp
        .cfi_def_cfa_offset 16
        .cfi_offset %rbp, -16
        movq    %rsp, %rbp
        .cfi_def_cfa_register %rbp
        pushq   %r14
        andq    $-16, %rsp
        subq    $16, %rsp
        .cfi_offset %r14, -24